requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_client]
pymongo
pydantic-settings
jinja2
